requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
rich>=13.7.0
win10toast>=0.9; sys_platform == "win32"
playwright>=1.45.0
python-dotenv>=1.0.1
geopy>=2.3.0